flask_app/requirements.txt
alembic==1.4.2
asgiref==3.2.3
astroid==2.3.3
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
pycparser==2.20
pygame==1.9.6
pylint==2.4.4
pyparsing==2.4.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.12.0
SQLAlchemy==1.3.15
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.9
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1